function evaluation implemented (still in test phase)

+++ GeoEngine/AST/Node_Expression/Binary_Expression/Booleans/Boolean_Expression/Equals_To.cs
@@ -8,6 +8,6 @@ public class EqualsTo : BooleanExpression
     {
         LeftNode.Evaluate();
         RightNode.Evaluate();
-        Value = LeftNode.Value.ToString() == RightNode.Value.ToString() ? 1 : 0;
+        Value = LeftNode.Value.ToString() == RightNode.Value.ToString() ? (double)1 : (double)0;
     }
 }
